# apps/backend/core-api/docs/swagger.yaml
definitions:
...
  core-api_internal_handler_event.EventResponse:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  core-api_internal_handler_eventconfig.EventCertificateConfigResponse:
    properties:
      academic_institution_font_family_id:
        type: integer
      academic_institution_font_weight:
        type: integer
      academic_institution_pos_x:
        type: number
      academic_institution_pos_y:
        type: number
      base_certificate_presigned_url:
        type: string
      base_certificate_storage_key:
        type: string
      certificate_subtitle_font_family_id:
        type: integer
      certificate_subtitle_font_weight:
        type: integer
      certificate_subtitle_pos_x:
        type: number
      certificate_subtitle_pos_y:
        type: number
      certificate_title_font_family_id:
        type: integer
      certificate_title_font_weight:
        type: integer
      certificate_title_pos_x:
        type: number
      certificate_title_pos_y:
        type: number
      created_at:
        type: string
      event_id:
        type: string
      event_name_font_family_id:
        type: integer
      event_name_font_weight:
        type: integer
      event_name_pos_x:
        type: number
      event_name_pos_y:
        type: number
      id:
        type: string
      is_published:
        type: boolean
      mint_readiness:
        $ref: '#/definitions/core-api_internal_handler_eventconfig.MintReadinessInfo'
      name_font_family_id:
        type: integer
      name_font_weight:
        type: integer
      name_pos_x:
        type: number
      name_pos_y:
        type: number
      updated_at:
        type: string
    required:
    - base_certificate_presigned_url
    - base_certificate_storage_key
    - created_at
    - event_id
    - id
    - is_published
    - name_pos_x
    - name_pos_y
    - updated_at
    type: object
  core-api_internal_handler_eventconfig.MintReadinessInfo:
    properties:
      all_issuers_have_signed:
        type: boolean
      certificate_contract_address:
        type: string
      has_certificate_config:
        type: boolean
      has_certificate_contract:
        type: boolean
      is_ready:
        type: boolean
      missing_requirements:
        items:
          type: string
        type: array
      signed_issuers_count:
        type: integer
      total_issuers_count:
        type: integer
    required:
    - all_issuers_have_signed
    - has_certificate_config
    - has_certificate_contract
    - is_ready
    - signed_issuers_count
    - total_issuers_count
    type: object
...
  customerror.ErrResponse:
    description: Response for the client to sign to register
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  entity.Event:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  entity.EventRegistrationInvitation:
    properties:
      academic_institution:
        type: string
      accepted_at:
        type: string
      cancelled_at:
        type: string
      code:
        type: string
      created_at:
        type: string
      email:
        type: string
      event_id:
        type: string
      first_name:
        type: string
      id:
        type: string
      inbox_message_id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      valid_until:
        type: string
    required:
    - accepted_at
    - created_at
    - event_id
    - id
    - inbox_message_id
    - updated_at
    type: object
  entity.EventStatus:
    enum:
    - active
    - inactive
    - closed
    type: string
    x-enum-varnames:
    - EventStatusActive
    - EventStatusInactive
    - EventStatusClosed
  entity.EventType:
    enum:
    - private
    - invite
    type: string
    x-enum-varnames:
    - EventTypePrivate
    - EventTypeInvite
  entity.InboxMessage:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      fallback_message_content:
        type: string
      hidden_at:
        type: string
      id:
        type: string
      is_read:
        type: integer
      message_content:
        type: string
      message_type:
        $ref: '#/definitions/entity.InboxMessageType'
      receiver_credential_id:
        type: string
      receiver_email:
        type: string
      receiver_wallet_address:
        type: string
      sender_credential_id:
        type: string
      updated_at:
        type: string
    required:
    - created_at
    - id
    - is_read
    - message_content
    - message_type
    - updated_at
    type: object
  entity.InboxMessageType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - InboxMessageTypeGeneral
    - InboxMessageTypeEventRegistrationInvitation
    - InboxMessageTypeEventCertificateInvitation
...
  event.GetEventListResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/entity.Event'
        type: array
    required:
    - events
    type: object
...
  event_registration.GetEventRegistrationInvitationByUserAndEventResponse:
    properties:
      inbox:
        $ref: '#/definitions/entity.InboxMessage'
      registration_invitation:
        $ref: '#/definitions/entity.EventRegistrationInvitation'
    type: object
...
  eventconfig.CreateEventRegistrationConfigRequest:
    properties:
      academic_email_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      academic_institution_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      address_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      bio_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      email_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      final_call_for_registration:
        type: string
      first_name_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      last_name_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      phone_number_requirement_status:
        enum:
        - 0
        - 1
        - 2
        type: integer
      registration_password:
        minLength: 8
        type: string
    required:
    - academic_email_requirement_status
    - academic_institution_requirement_status
    - address_requirement_status
    - bio_requirement_status
    - email_requirement_status
    - first_name_requirement_status
    - last_name_requirement_status
    - phone_number_requirement_status
    type: object
  eventconfig.EventRegistrationConfigRequirementStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - EventRegistrationConfigRequirementStatusNotRequired
    - EventRegistrationConfigRequirementStatusRequired
    - EventRegistrationConfigRequirementStatusOptional
  eventconfig.EventRegistrationConfigViewModel:
    properties:
      academic_email_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      academic_institution_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      address_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      bio_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      created_at:
        type: string
      email_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      event_id:
        type: string
      final_call_for_registration:
        type: string
      first_name_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      id:
        type: string
      last_name_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      phone_number_requirement_status:
        $ref: '#/definitions/eventconfig.EventRegistrationConfigRequirementStatus'
      updated_at:
        type: string
    required:
    - academic_email_requirement_status
    - academic_institution_requirement_status
    - address_requirement_status
    - bio_requirement_status
    - created_at
    - email_requirement_status
    - event_id
    - first_name_requirement_status
    - id
    - last_name_requirement_status
    - phone_number_requirement_status
    - updated_at
    type: object
...
  eventconfig.ToggleCertificatePublishedRequest:
    properties:
      is_published:
        type: boolean
    required:
    - is_published
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: DECM (Decentralized Event Management) platform API for NFT ticketing,
    digital credentials, and academic identity verification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DECM Core
  version: "1.0"
paths:
...
  /api/v1/event-registration/my/{event_id}:
    get:
      consumes:
      - application/json
      description: Get event registration invitation of user and by event id
      operationId: get-event-registration-invitation-by-user-and-event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event_registration.GetEventRegistrationInvitationByUserAndEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get event registration invitation of user and by event id
      tags:
      - Event Registration Invitation
...
  /api/v1/events:
    get:
      description: Get events list
      operationId: get-events-list
      parameters:
      - default: true
        description: Include active events
        in: query
        name: include_active_events
        type: boolean
      - default: true
        description: Include inactive events
        in: query
        name: include_inactive_events
        type: boolean
      - default: false
        description: Include closed events
        in: query
        name: include_closed_events
        type: boolean
      - default: false
        description: Only user joined events
        in: query
        name: only_user_joined_events
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.GetEventListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get events list
      tags:
      - Event
    post:
      consumes:
      - multipart/form-data
      description: Create a new event with banner image upload. Chain ID is automatically
        set from environment configuration.
      operationId: create-event
      parameters:
      - description: Event name
        in: formData
        name: name
        required: true
        type: string
      - description: Event short description
        in: formData
        name: short_description
        required: true
        type: string
      - description: Event description
        in: formData
        name: description
        required: true
        type: string
      - description: Start date (RFC3339 format)
        in: formData
        name: start_date
        required: true
        type: string
      - description: End date (RFC3339 format)
        in: formData
        name: end_date
        required: true
        type: string
      - description: Number of seats
        in: formData
        name: seats_count
        required: true
        type: integer
      - description: Contact number
        in: formData
        name: contact_number
        required: true
        type: string
      - description: Contact address
        in: formData
        name: contact_address
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Google map query
        in: formData
        name: google_map_query
        required: true
        type: string
      - description: Event banner image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: banner
        required: true
        type: file
      - description: Event icon image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: icon
        required: true
        type: file
      - description: Host password
        in: formData
        name: host_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core-api_internal_handler_event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Create a new event
      tags:
      - Event
...
  /api/v1/events/{event_id}/config/certificate/published:
    patch:
      consumes:
      - application/json
      description: Toggle the published status of event certificate configuration.
        When setting to published, inbox messages will be created for all certificate
        receivers. Only accessible by verified organizers.
      operationId: toggle-certificate-published
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Toggle published request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/eventconfig.ToggleCertificatePublishedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core-api_internal_handler_eventconfig.EventCertificateConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Toggle certificate published status
  /api/v1/events/{event_id}/config/certificate/template:
    get:
      description: Proxy the base certificate SVG template from storage so the frontend
        can access it without CORS restrictions from presigned URLs. Accessible by
        the event owner or issuers assigned to the event.
      operationId: get-event-certificate-template
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: SVG content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get event certificate template SVG
...
  /api/v1/events/{event_id}/registration-config:
    delete:
      consumes:
      - application/json
      description: Delete the event registration configuration for an event
      operationId: delete-event-registration-config
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Delete event registration config
    post:
      consumes:
      - application/json
      description: Create a new event registration configuration for an event
      operationId: create-event-registration-config
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Event registration config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/eventconfig.CreateEventRegistrationConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eventconfig.EventRegistrationConfigViewModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Create event registration config
...
swagger: "2.0"